Refactor code for the overview module
