+++ src/main/java/life/qbic/portal/interfaces/ProjectOverviewI.java
@@ -1,0 +1,7 @@
+package life.qbic.portal.interfaces;
+
+/**
+ * Created by sven1103 on 8/12/16.
+ */
+public interface ProjectOverviewI {
+}
