+++ src/test/java/life.qbic.portal/TestSuite.java
@@ -1,0 +1,7 @@
+package life.qbic.portal;
+
+/**
+ * Created by sven1103 on 8/12/16.
+ */
+public class TestSuite {
+}
